+++ CHANGELOG.md
@@ -7,10 +7,15 @@ The format is based on
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) this project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Multi-cursor support
+
 ## [1.0.0] - 2018-05-21
 ### Added
 - Now respects VSCode's `editor.wordSeparators` configuration
-  ([PR #2](https://github.com/sgryjp/japanese-word-handler/pull/2), contribution by [@tekezo](https://github.com/tekezo))
+  ([PR #2](https://github.com/sgryjp/japanese-word-handler/pull/2), contribution
+  by [@tekezo](https://github.com/tekezo))
 
 ### Fixed
 - Moving cursor with Ctrl+Left over whitespaces at start of a document fails